requirements.txt
async-timeout==4.0.2
click==8.1.3
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
redis==4.4.2
Werkzeug==2.2.2
python-dotenv==0.21.0
gunicorn==20.1.0
